feat: Add BadRequestException, NotFoundException and ValidationException.

+++ InventoryManagment.Application/Features/Products/Handlers/Commands/DeleteProductCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using InventoryManagment.Application.Exceptions;
 using InventoryManagment.Application.Features.Products.Requests.Commands;
 using InventoryManagment.Application.Persistence.Contracts;
 using MediatR;
@@ -17,6 +18,9 @@ namespace InventoryManagment.Application.Features.Products.Handlers.Commands
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
             var product = await _productRepository.GetAsync(request.Id);
+
+            if (product == null)
+                throw new NotFoundException(nameof(product), request.Id);
 
             await _productRepository.DeleteAsync(product);
